requirements.txt
mysql-connector-python==9.4.0
python-dotenv==1.1.1
streamlit==1.49.1
pandas==2.3.2
babel==2.17.0
